requirements.txt
Flask
psycopg2
jinja2
python-dotenv
